speaking_buddy/pyproject.toml
[project]
name = "speaking-buddy"
version = "0.1.0"
description = "Pronunciation learning tool for Luxembourgish"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.32.0",
    "streamlit-audiorecorder>=0.0.5",
    "pydub>=0.25.1",
    "scipy>=1.12.0",
    "numpy>=1.26.0,<2.0",
    "librosa>=0.10.2",
    "soundfile>=0.12.1",
    "requests>=2.31.0",
    "numba>=0.60.0",
    "praat-parselmouth>=0.4.3",
]
